+++ src/test/integration/index.ts
@@ -3,7 +3,7 @@ import Mocha from 'mocha';
 import * as fs from 'fs';
 
 export function run(): Promise<void> {
-  const mocha = new Mocha({ ui: 'bdd', color: true, timeout: 30000 });
+  const mocha = new Mocha({ ui: 'tdd', color: true, timeout: 30000 });
   const testsRoot = path.resolve(__dirname);
 
   return new Promise((resolve, reject) => {